backend/requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1

# OpenAI Chatbot
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2

# Google Gemini (alternativa)
google-generativeai==0.3.2

# Firebase Push Notifications
firebase-admin==6.3.0

# Environment Variables
python-decouple==3.8
python-dotenv==1.0.0

# Image Upload
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.10

# Celery (tareas programadas)
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Utils
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
flake8==7.0.0
black==23.12.1
